// src/models/TaskStateModel.ts
import type { TaskModel } from './TaskModel';

export type TaskStateModel = {
  tasks: TaskModel[];
  secondsRemaining: number;
  formattedTimeRemaining: string;
  activeTask: TaskModel | null;
  currentCycle: number; // 1 to 8
  config: {
    workTime: number;
    shortBreakTime: number;
    longBreakTime: number;
  };
};